appdata: Add remote links to an icon
There is a real icon which is used when the app is installed, but
that is in the extra data, which we don't want to redistribute.
So, for the appdata we instead use a remote icon, as such linking
should be ok (we're not redistributing the file).

The currently used icon is the on from the android client
from the play.google.com site.