Define supported & required elements
The internet field is a	bit tricky, I wanted to	put 'first-run'	since
you can	technically launch it, log in, download	music, and never have
it connect to the internet again.

However	the appstream spec defines first-run as:
"Uses the internet the first time the application is run, but not
normally afterwards."

Which is not reflective	of the Spotify experience. Most	users will
connect after first-run.

We can't put it	in the recommends element either since the spec	says
for that:
"...then this indicates that the app can work without internet, but the
experience will be degraded."

Which is also not accurate since it has	to log in at least once, so just
put always.

https://support.spotify.com/us/article/audio-quality